requirements.txt
# Core dependencies
git+https://github.com/DS4SD/docling.git@v2.19.0

# OpenAI and utilities
python-dotenv==1.0.1
openai==1.12.0
pydantic==2.6.1
tiktoken==0.6.0

# Database
psycopg2-binary==2.9.9
pgvector==0.2.4
asyncpg==0.29.0

# Document processing
PyPDF2==3.0.1
pdfplumber==0.10.2
python-docx>=1.1.2,<2.0.0

# Data analysis
numpy==1.26.3
pandas==2.2.0

# CLI interface
rich==13.7.0

# Optional dependencies
# Uncomment if you need PgVector search functionality
# sqlalchemy==2.0.27
# sqlalchemy-utils==0.41.1
